Make use of the OSS SWIM implementation (#746)
* !swim moving to oss  swim implementation; compiles

* All actor refs now contain an UniqueNode, even local ones; this vastly
simplifies many patterns and removes a lot of fatalErrors in our code;
Eases SWIM integration since it assumes it is working with Nodes

* more peer usage

* working on unlocking tests

* unlock test_swim_shouldIncreaseProbeInterval_whenHighMultiplier

* connect tests dont make sense anymore; refs manage this automatically

* unlock test_swim_shouldPingSpecificMemberWhenRequested

* =deadletters should use path in actor/path metadata

* fix wrong assumption how a remote node is detected

* debugging, unlocks test_swim_shouldSendGossipInAck

* compiling against 0.1.0 SWIM

* the test was wrong

remove wrong test; as known since #719

fix wire envelope serialization on 5.3

logging handling

* change docs to be more actor specific and point to SWIM for details

* =deadletters improve how we print dead letters

confirming dead nodes now works always

use main branch of cluster membership for now

* =test,fix,singleton fix singleton resolve code after address changes

* test fixups & swift 5.2.5

* =fix serialization test when 5.3, linux and working getting manifests exists

* fixup compilation

* review adjustments

* Update Docs/clustering.adoc

+++ Sources/DistributedActors/ProcessIsolated/ProcessIsolated.swift
@@ -498,6 +498,7 @@ public enum ProcessSpawnError: Error {
 
 extension UniqueNode {
     // Very naive parsing method for nodes; but good enough for our simple needs here.
+    // FIXME: should become throwing, nowadays faults are fatal
     internal static func parse(_ string: String) -> UniqueNode {
         var s: Substring = string[...]
         s = s.dropFirst("sact://".count)
@@ -509,7 +510,7 @@ extension UniqueNode {
         let _nid = String(s.prefix(while: { $0 != "@" }))
         s = s.dropFirst(_nid.count)
         s = s.dropFirst(":".count)
-        let nid = UniqueNodeID(UInt32(_nid)!)
+        let nid = UniqueNodeID(UInt64(_nid)!)
 
         let host = String(s.prefix(while: { $0 != ":" }))
         s = s.dropFirst(host.count)
